requirements.txt
openai
streamlit
pypdf
requests
python-dotenv
